requirements.txt
requests==2.25.0
pyyaml==5.3.1
pygments==2.7.2
mkdocs==1.1.2